+++ .vscode/settings.json
@@ -29,6 +29,7 @@
 		"authly",
 		"bytewise",
 		"cloudly",
+		"crockford",
 		"cryptly",
 		"Encrypter",
 		"Encrypters",
