+++ Authenticator/index.ts
@@ -1,9 +1,11 @@
-import * as Base16 from "./Base16"
-import * as Base32 from "./Base32"
-import { Signer } from "./Signer"
+import { Base16 } from "../Base16"
+import { Base32 } from "../Base32"
+import { Signer } from "../Signer"
 
-export namespace authenticator {
-	export async function generate(key: string, time: number, length = 6): Promise<string> {
+export type Authenticator = string
+
+export namespace Authenticator {
+	export async function generate(key: string, time: number, length = 6): Promise<Authenticator> {
 		const hash = await Signer.create("HMAC", "SHA-1", new TextEncoder().encode(key)).sign(
 			Base16.decode(
 				Math.floor(time / 1000 / 30)
@@ -21,14 +23,10 @@ export namespace authenticator {
 		// (value % 10^digits).padstart(digits, "0")
 		return (value % +"1".padEnd(length + 1, "0")).toString().padStart(length, "0")
 	}
-	export async function generateRecoveryCodes(key: string, times: number[]): Promise<string[]> {
-		const result = Array(times.length)
-		for (let index = 0; index < times.length; index++) {
-			result[index] = await generate(key, times[index], 8)
-		}
-		return result
+	export function generateRecoveryCodes(key: string, times: number[]): Promise<Authenticator[]> {
+		return Promise.all(times.map(async time => await generate(key, time, 8)))
 	}
-	export function toQrCode(key: string, issuer: string, username: string): string {
+	export function toQRCode(key: string, issuer: string, username: string): string {
 		return `otpauth://totp/${issuer}:${username}?secret=${Base32.encode(key)}&issuer=${issuer}`
 	}
 }
