+++ Base16/index.spec.ts
@@ -1,4 +1,4 @@
-import { cryptly } from "./index"
+import { cryptly } from "../index"
 
 describe("Base16", () => {
 	it("encode standard 1", () =>
@@ -10,19 +10,19 @@ describe("Base16", () => {
 	it("encode standard 4", () =>
 		expect(cryptly.Base16.encode("any carnal pleasur")).toEqual("616e79206361726e616c20706c6561737572"))
 	it("decode standard 1", () =>
-		expect(new cryptly.TextDecoder().decode(cryptly.Base16.decode("5468697320697320746865206461746120282a29"))).toEqual(
+		expect(new TextDecoder().decode(cryptly.Base16.decode("5468697320697320746865206461746120282a29"))).toEqual(
 			"This is the data (*)"
 		))
 	it("decode standard 2", () =>
-		expect(new cryptly.TextDecoder().decode(cryptly.Base16.decode("616e79206361726e616c20706c6561737572652e"))).toEqual(
+		expect(new TextDecoder().decode(cryptly.Base16.decode("616e79206361726e616c20706c6561737572652e"))).toEqual(
 			"any carnal pleasure."
 		))
 	it("decode standard 3", () =>
-		expect(new cryptly.TextDecoder().decode(cryptly.Base16.decode("616e79206361726e616c20706c656173757265"))).toEqual(
+		expect(new TextDecoder().decode(cryptly.Base16.decode("616e79206361726e616c20706c656173757265"))).toEqual(
 			"any carnal pleasure"
 		))
 	it("decode standard 4", () =>
-		expect(new cryptly.TextDecoder().decode(cryptly.Base16.decode("616e79206361726e616c20706c6561737572"))).toEqual(
+		expect(new TextDecoder().decode(cryptly.Base16.decode("616e79206361726e616c20706c6561737572"))).toEqual(
 			"any carnal pleasur"
 		))
 	it("xor", () =>
