+++ Encrypter/Aes/index.ts
@@ -1,7 +1,5 @@
-import * as Base64 from "../../Base64"
+import { Base64 } from "../../Base64"
 import { crypto } from "../../crypto"
-import { TextDecoder } from "../../TextDecoder"
-import { TextEncoder } from "../../TextEncoder"
 import { Encrypted as AesEncrypted } from "./Encrypted"
 
 export class Aes {
