+++ Signer/Base.ts
@@ -1,5 +1,4 @@
-import * as Base64 from "../Base64"
-import { TextEncoder } from "../TextEncoder"
+import { Base64 } from "../Base64"
 
 export abstract class Base {
 	async sign(data: Uint8Array): Promise<Uint8Array>
