--- TextEncoder.ts
@@ -1,6 +1,0 @@
-export class TextEncoder {
-	readonly encoding = "utf-8"
-	encode(data: string): Uint8Array {
-		return Uint8Array.from(unescape(encodeURIComponent(data)).split(""), c => c.charCodeAt(0))
-	}
-}
