--- Tuple.ts
@@ -1,2 +1,0 @@
-// Workaround for ESLint parsing issues with Tuples
-export type Tuple2<T1, T2> = [T1, T2]
